mobile

+++ mobile/mobile.go
@@ -1,5 +1,3 @@
-// +build android ios darwin,arm darwin,arm64
-
 package inovation
 
 import (
@@ -40,6 +38,6 @@ func UpdateTouchesOnAndroid(action int, id int, x, y int) {
 	mobile.UpdateTouchesOnAndroid(action, id, x, y)
 }
 
-func UpdateTouchesOnIOS(phase int, ptr int, x, y int) {
+func UpdateTouchesOnIOS(phase int, ptr int64, x, y int) {
 	mobile.UpdateTouchesOnIOS(phase, ptr, x, y)
 }
